Refactor reminders so we show content whilst the data is loading
