+++ src/client/modules/Reminders/constants.js
@@ -2,3 +2,17 @@ export const settings = {
   OFF: 'Off',
   ON: 'On',
 }
+
+export const sortOptions = [
+  {
+    name: 'Most recent',
+    value: '-created_on',
+  },
+  {
+    name: 'Oldest',
+    value: 'created_on',
+  },
+]
+
+export const maxItemsToPaginate = 10000
+export const itemsPerPage = 10
